make the qwt5 dependency optional
The only widget that depends on it is the Compass widget in
artificial_horizon/. Unfortunately, this is a pretty common
widget. We would in fine need to reimplement it properly
without qwt.

In the meantime, make sure we can build without QWT
installed. The qt4 and qt5 dev files for QWT cannot
be installed in parallel, which causes problems with
e.g. gazebo dev files.